+++ app/controllers/ClaimAnyOtherExpenseController.scala
@@ -18,33 +18,39 @@ package controllers
 
 import connectors.DataCacheConnector
 import controllers.actions._
-import forms.CovidHomeWorkingFormProvider
-import identifiers.CovidHomeWorkingId
+import forms.OnlyWorkingFromHomeExpensesFormProvider
+import identifiers.{ClaimantId, OnlyWorkingFromHomeExpensesId}
 import javax.inject.Inject
+import models.Claimant
 import play.api.data.Form
 import play.api.i18n.I18nSupport
+import play.api.libs.json.{JsBoolean, JsString, JsValue}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
+import uk.gov.hmrc.http.cache.client.CacheMap
 import uk.gov.hmrc.play.bootstrap.controller.FrontendBaseController
 import utils.{Enumerable, Navigator, UserAnswers}
-import views.html.CovidHomeWorkingView
+import views.html.ClaimAnyOtherExpenseView
 
 import scala.concurrent.{ExecutionContext, Future}
 
-class CovidHomeWorkingController @Inject()(
+
+
+class ClaimAnyOtherExpenseController @Inject()(
                                      dataCacheConnector: DataCacheConnector,
                                      navigator: Navigator,
                                      getData: DataRetrievalAction,
                                      flowEnabled: WorkingFromHomeEnabledAction,
-                                     formProvider: CovidHomeWorkingFormProvider,
+                                     formProvider: OnlyWorkingFromHomeExpensesFormProvider,
                                      val controllerComponents: MessagesControllerComponents,
-                                     view: CovidHomeWorkingView
+                                     view: ClaimAnyOtherExpenseView
                                    )(implicit ec: ExecutionContext) extends FrontendBaseController with I18nSupport with Enumerable.Implicits {
 
   val form: Form[Boolean] = formProvider()
 
+
   def onPageLoad: Action[AnyContent] = (flowEnabled andThen getData) {
     implicit request =>
-      val preparedForm = request.userAnswers.flatMap(_.covidHomeWorking) match {
+      val preparedForm = request.userAnswers.flatMap(_.onlyWorkingFromHomeExpenses) match {
         case None => form
         case Some(value) => form.fill(value)
       }
@@ -56,10 +62,20 @@ class CovidHomeWorkingController @Inject()(
       form.bindFromRequest().fold(
         (formWithErrors: Form[_]) =>
           Future.successful(BadRequest(view(formWithErrors))),
-        value =>
-          dataCacheConnector.save[Boolean](request.sessionId, CovidHomeWorkingId, value).map(cacheMap =>
-            Redirect(navigator.nextPage(CovidHomeWorkingId)(new UserAnswers(cacheMap)))
+        value => {
+
+          val cacheMap = CacheMap(request.sessionId, Map[String, JsValue](
+            ClaimantId.toString                     -> JsString( Claimant.You.string),
+            OnlyWorkingFromHomeExpensesId.toString  -> JsBoolean(value)
+          ))
+
+          dataCacheConnector.save(cacheMap).map(cacheMap =>
+            Redirect(navigator.nextPage(OnlyWorkingFromHomeExpensesId)(new UserAnswers(cacheMap)))
           )
+
+        }
+
       )
   }
+
 }
