+++ app/views/ClaimAnyOtherExpenseView.scala.html
@@ -33,7 +33,7 @@
 
     @back_link()
 
-    @formHelper(action = OnlyWorkingFromHomeExpensesController.onSubmit(), 'autoComplete -> "off") {
+    @formHelper(action = ClaimAnyOtherExpenseController.onSubmit(), 'autoComplete -> "off") {
 
         @components.error_summary(form.errors)
