+++ app/views/WfhDueToCovidView.scala.html
@@ -33,7 +33,7 @@
 
     @back_link()
 
-    @formHelper(action = CovidHomeWorkingController.onSubmit(), 'autoComplete -> "off") {
+    @formHelper(action = WfhDueToCovidController.onSubmit(), 'autoComplete -> "off") {
 
         @components.error_summary(form.errors)
